pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[project]
name = "wireviz"
version = "0.5.4"
description = "Easily document cables and wiring harnesses"
readme = "docs/README.md"
requires-python = ">=3.8"
license = { text = "GPL-3.0-only" }
authors = [{ name = "Daniel Rojas" }]
dependencies = [
    "click",
    "graphviz",
    "pillow",
    "pyyaml",
    "tabulate",
    "jinja2",
    "weasyprint",
    "pydantic<2",
]

[project.scripts]
filare = "wireviz.wv_cli:cli"
filare-qty = "wireviz.models.harness_quantity:qty_multipliers"
wireviz = "wireviz.wv_cli:cli"
wireviz-qty = "wireviz.models.harness_quantity:qty_multipliers"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "coverage[toml]>=7.6.0",
    "isort>=5.12.0",
]

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "coverage[toml]>=7.6.0",
    "isort>=5.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--maxfail=1 -ra --cov=wireviz --cov-report=term-missing --cov-report=html:report/coverage"
markers = [
    "examples: slow example generation",
    "tutorial: slow tutorial generation",
]

[tool.coverage.run]
branch = true
source = ["src/wireviz"]
omit = [
    "src/wireviz/render/pdf.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "report/coverage"

[tool.isort]
profile = "black"
